Move boilerplate & fact sheet above the fold; drop Coby from spokespeople
Address review feedback on the /press page:
- Extract the boilerplate and fact sheet into a new About section placed at
  the top of the page (and nav) so journalists see them first.
- PressKit now only carries logos and imagery.
- Remove Coby Cotton from the spokespeople list.
